move rfsoc_server to server
