+++ tests/test_server.py
@@ -9,7 +9,7 @@ import qibosoq
 from qibosoq.components.base import Parameter
 from qibosoq.components.pulses import Rectangular
 from qibosoq.log import define_loggers
-from qibosoq.rfsoc_server import execute_program, load_pulses
+from qibosoq.server import execute_program, load_pulses
 
 qibosoq.configuration.MAIN_LOGGER_FILE = "/tmp/test_log_rfsoc.log"
 qibosoq.configuration.PROGRAM_LOGGER_FILE = "/tmp/test_log2_rfsoc.log"
